requirements.txt
pyTelegramBotAPI
Pillow
qrcode
pyshorteners
edge-tts